refactor(shell): delete the dead pre-worker shell UI elements
The shell still carried a cluster of Lit custom elements and one job
queue left over from the charm-list shell that preceded the worker
cutover. Every one of them was registered at browser startup and none of
them was ever rendered: the ACL editor was rendered only by a view that
no longer exists, the piece-link element and the piece view that went
with it lost their last render site at the same time, the spinner was
rendered only by that piece view, and the work queue was the command
processor of an architecture the shell no longer uses. Dead elements are
worse than dead functions, because registering a custom element is a
side effect: the code runs on every page load, and anyone reading the
component index has to work out from scratch whether an element is live
before touching it.

Delete ACLView, PieceLink, _PieceView, Spinner, and the WorkQueue class,
and drop the corresponding entries from the view and component indexes.

The flex elements needed more care than a straight deletion. Three
elements were defined together: XFlexElement did the layout, HBoxElement
rendered an XFlexElement in its shadow root, and VBoxElement was
HBoxElement with the direction flipped to vertical. Only <v-box> is
still used, in the recovery-phrase panel of the login view, so this
change collapses the three-element chain into a single standalone
element and renames the file after it.

The new element reproduces what the old chain actually rendered, which
is not quite what the old stylesheet said. The host takes over the block
display and the full width and height that XFlexElement had, and the
shadow root holds one column-direction flex container. The old
`::slotted(*) { flex: 1 }` rule is deliberately not carried over: in the
old chain that selector matched the intermediate slot element handed
down from HBoxElement's shadow root rather than the caller's children,
and because a slot defaults to `display: contents`, the rule never
generated a box and so never applied to anything. Reintroducing it
against the real children would be a behavior change rather than a
faithful port. The `center` property is dropped for want of any caller.

Rendering both the old chain and the new element in a browser and
comparing the measured geometry of the recovery-phrase panel gives
identical results, and the same panel measured in the running shell
stacks the text area above the copy button at full container width, as
before.

Two loose ends go with it. The header view carried a side-effect import
of the flex module without rendering any of its elements, while the
login view rendered <v-box> and relied on the component index to have
defined it; the import now sits with the view that uses it. The body
view's stylesheet had a `v-box` rule and no <v-box> in its template. A
comment in the navigation module listing the ways navigation can start
named the deleted piece-link element, and now names the function such a
control calls.
