+++ packages/shell/src/components/index.ts
@@ -1,7 +1,5 @@
 export * from "./Button.ts";
-export * from "./PieceLink.ts";
 export * from "./PieceList.ts";
 export * from "./CFLogo.ts";
-export * from "./Flex.ts";
-export * from "./Spinner.ts";
+export * from "./VBox.ts";
 export * from "./icons.ts";
