+++ packages/shell/src/views/LoginView.ts
@@ -19,6 +19,7 @@ import {
   type StoredCredential,
 } from "../lib/credentials.ts";
 import { ROOT_KEY } from "../lib/root-key.ts";
+import "../components/VBox.ts";
 
 type AuthFlow = "register" | "login";
