+++ packages/shell/src/views/index.ts
@@ -1,4 +1,3 @@
-import "./ACLView.ts";
 import "./AppView.ts";
 import "./BodyView.ts";
 import "./DebuggerView.ts";
